Add all my files

+++ resources/views/products/index.blade.php
@@ -4,12 +4,18 @@
     {{"All products"}}
 @endsection
 @section("content")
+
 <section class="productCard_btnSection">
     <button class="productCard_button" onclick="window.location.href='/'"> Home </button>
     <button class="productCard_button" onclick="window.location.href='/product/create'"> Upload product </button>
-    <button class="productCard_button" onclick="window.location.href='/product/destroy'"> verwijder product </button>
-    <button class="productCard_button" onclick="window.location.href='/dashboard'"> Ga naar je profiel </button>
+    @if(Auth::user()->role === "Admin")
+        <button class="productCard_button" onclick="window.location.href='/product/destroy'"> verwijder product </button>
+    @endif
     <button class="productCard_button" onclick="window.location.href='/user'"> Bekijk profielen </button>
+    <form method="POST" action="/logout">
+        @csrf
+        <button type="submit" class="productCard_button_uitloggen">Uitloggen</button>
+     </form>
 </section>
 <input class="typeBar" type="text" id="myInput" onkeyup="searchbar()" placeholder="Search for names.." title="Type in a name">
 
@@ -21,4 +27,8 @@
         @include("products.components.productCard--index")
     @endforeach
 </ul>
+
+
 @endsection
+
+
